Nodecontroller
