+++ node/tests/reconfigure.rs
@@ -6,12 +6,12 @@ use config::{Committee, Parameters};
 use consensus::ConsensusOutput;
 use crypto::{
     ed25519::{Ed25519KeyPair, Ed25519PublicKey},
-    traits::{KeyPair, Signer, VerifyingKey},
+    traits::{KeyPair, VerifyingKey},
 };
-use executor::{ExecutionIndices, ExecutionState, ExecutionStateError, ExecutorOutput};
+use executor::{ExecutionIndices, ExecutionState, ExecutionStateError};
 use futures::future::join_all;
 use network::{PrimaryNetwork, PrimaryToWorkerNetwork};
-use node::{Node, NodeStorage};
+use node::{Node, NodeController, NodeStorage};
 use primary::PrimaryWorkerMessage;
 use prometheus::Registry;
 use std::{
@@ -121,97 +121,6 @@ impl ExecutionStateError for SimpleExecutionError {
     }
 }
 
-async fn run_node<Keys, PublicKey, State>(
-    keypair: Keys,
-    committee: &Committee<PublicKey>,
-    execution_state: Arc<State>,
-    parameters: Parameters,
-    mut rx_reconfigure: Receiver<(Keys, Committee<PublicKey>)>,
-    tx_output: Sender<ExecutorOutput<State>>,
-) where
-    PublicKey: VerifyingKey,
-    State: ExecutionState + Send + Sync + 'static,
-    State::Outcome: Send + 'static,
-    State::Error: Debug,
-    Keys: KeyPair<PubKey = PublicKey> + Signer<PublicKey::Sig> + Send + 'static,
-{
-    let mut keypair = keypair;
-    let mut name = keypair.public().clone();
-    let mut committee = committee.clone();
-
-    let mut handles = Vec::new();
-    let mut primary_network = PrimaryNetwork::default();
-    let mut worker_network = PrimaryToWorkerNetwork::default();
-
-    // Listen for new committees.
-    loop {
-        // Get a fresh store for the new epoch.
-        let store = NodeStorage::reopen(test_utils::temp_dir());
-
-        // Restart the relevant components.
-        let primary_handles = Node::spawn_primary(
-            keypair,
-            Arc::new(ArcSwap::new(Arc::new(committee.clone()))),
-            &store,
-            parameters.clone(),
-            /* consensus */ true,
-            execution_state.clone(),
-            tx_output.clone(),
-            &Registry::new(),
-        )
-        .await
-        .unwrap();
-
-        let worker_handles = Node::spawn_workers(
-            name.clone(),
-            /* worker_ids */ vec![0],
-            Arc::new(ArcSwap::new(Arc::new(committee.clone()))),
-            &store,
-            parameters.clone(),
-            &Registry::new(),
-        );
-
-        handles.extend(primary_handles);
-        handles.extend(worker_handles);
-
-        // Wait for a committee change.
-        let (new_keypair, new_committee) = match rx_reconfigure.recv().await {
-            Some(x) => x,
-            None => break,
-        };
-
-        // Shutdown all relevant components.
-        let address = committee
-            .primary(&name)
-            .expect("Our key is not in the committee")
-            .primary_to_primary;
-        let message = PrimaryMessage::<PublicKey>::Reconfigure(ReconfigureNotification::Shutdown);
-        let primary_cancel_handle = primary_network.send(address, &message).await;
-
-        let addresses = committee
-            .our_workers(&name)
-            .expect("Our key is not in the committee")
-            .into_iter()
-            .map(|x| x.primary_to_worker)
-            .collect();
-        let message =
-            PrimaryWorkerMessage::<PublicKey>::Reconfigure(ReconfigureNotification::Shutdown);
-        let worker_cancel_handles = worker_network.broadcast(addresses, &message).await;
-
-        // Ensure the message has been received.
-        primary_cancel_handle.await;
-        join_all(worker_cancel_handles).await;
-
-        // Wait for the components to shut down.
-        join_all(handles.drain(..)).await;
-
-        // Update the settings for the next epoch.
-        keypair = new_keypair;
-        name = keypair.public().clone();
-        committee = new_committee;
-    }
-}
-
 async fn run_client<PublicKey: VerifyingKey>(
     name: PublicKey,
     committee: Committee<PublicKey>,
@@ -279,9 +188,10 @@ async fn restart() {
         let execution_state = execution_state.clone();
         let parameters = parameters.clone();
         tokio::spawn(async move {
-            run_node(
+            NodeController::watch(
                 keypair,
                 &committee,
+                /* base_store_path */ test_utils::temp_dir(),
                 execution_state,
                 parameters,
                 rx_node_reconfigure,
@@ -356,6 +266,7 @@ async fn epoch_change() {
         ));
         states.push(execution_state.clone());
 
+        // Start a task that will broadcast the committee change signal.
         let name_clone = name.clone();
         tokio::spawn(async move {
             let mut primary_network = PrimaryNetwork::default();
